pyollama/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyollama"
version = "0.1.0"
description = "A complete Python implementation of Ollama for running large language models locally"
authors = [
    {name = "AI Assistant", email = "ai@example.com"},
]
dependencies = [
    "llama-cpp-python[cuda]",
    "huggingface-hub",
    "numpy",
    "click",
    "fastapi",
    "uvicorn",
    "pydantic",
    "requests",
    "tqdm",
    "psutil",
    "torch",
    "transformers",
    "accelerate",
    "sentencepiece",
    "protobuf",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
pyollama = "pyollama.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
